+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 ### Release
 
+**_Version 0.21.6_**
+
+- Hook change named export to default export
+
+**_Version 0.21.5_**
+
+- storybook button icon label problem fix
+
+**_Version 0.21.4_**
+
+- İnfinite table export
+
 **_Version 0.21.3_**
 
 - fix storybook button label problem